Add a social status
All Romans are part of family. That family has one of two social
statuses: Patrician or Plebian. Plebians were a larger proportion of the
population we weight the random generation 70/30 towards Plebians
instead of using a "fair" 50/50 roll.

A nomen is now associated with a family. For higher historical accuracy,
we generate random families based on their social status from lists of
historical plebian/patrician families.